+++ tests/testthat/ref/invisible_returned_obj.R
@@ -1,0 +1,4 @@
+my_fun <- function(x) {
+  return(invisible(x))
+}
+my_fun("do not print this")